requirements.txt
aiofiles
deta
hypercorn
jinja2
jsonify
fastapi
python-dotenv
